Cambiar estado de tareas y agregar Controller para las funciones. (file about deleted)
