+++ src/controllers/task.controller.js
@@ -1,0 +1,44 @@
+import Task from "../models/Task";
+
+export const renderTasks = async (req, res) => {
+    const tasks = await Task.find().lean();
+    res.render('index', { tasks: tasks });
+}
+
+export const createTask = async (req,res) => {
+    try {
+        const task = Task(req.body);
+        await task.save();
+        res.redirect('/');
+    } catch (error) {
+        console.log(error);
+    }
+}
+
+export const renderEditTask = async (req,res) => {
+    try {
+        const task = await Task.findById(req.params.id).lean();
+        res.render('edit', {task: task});
+    } catch (error) {
+        console.log(error);
+    }
+}
+
+export const editTask = async (req,res) => {
+    const result = await Task.findByIdAndUpdate(req.params.id, req.body);
+    res.redirect('/');
+}
+
+export const deleteTask = async (req,res) => {
+    const id = req.params.id;
+    await Task.findByIdAndDelete(id);
+    res.redirect('/');
+}
+
+export const taskToggleDone = async (req,res) => {
+    const id = req.params.id;
+    const task = await Task.findById(id);
+    task.done = !task.done;
+    await task.save();
+    res.redirect('/');
+}
