--- src/views/about.hbs
@@ -1,1 +1,0 @@
-<h1>about</h1>
