+++ src/views/edit.hbs
@@ -1,4 +1,4 @@
-<form action="/edit/{{task._id}}" method="POST">
+<form action="/editTask/{{task._id}}" method="POST">
     <input type="text" name="title" placeholder="Título" value="{{task.title}}">
     <textarea name="description" rows="3" placeholder="Agregue una descripción">{{task.description}}</textarea>
     <button>Actualizar</button>
