+++ src/views/partials/navbar.hbs
@@ -1,6 +1,5 @@
 <h1>Navbar</h1>
 <ul>
     <li><a href="/">index</a></li>
-    <li><a href="/about">about</a></li>
     <li><a href="/edit">edit</a></li>
 </ul>
