+++ src/views/partials/tasks/taskTable.hbs
@@ -14,9 +14,13 @@
                 <td>{{title}}</td>
                 <td>{{description}}</td>
                 <td>
-                    <button>Done</button>
-                    <a href="/delete/{{_id}}">Eliminar</a>
-                    <a href="/edit/{{_id}}">Edit</a>
+                    {{#if done}}
+                        <a href="/toggleDoneTask/{{_id}}">Por Hacer</a>
+                    {{else}}
+                        <a href="/toggleDoneTask/{{_id}}">Hecho</a>
+                    {{/if}}
+                    <a href="/deleteTask/{{_id}}">Eliminar</a>
+                    <a href="/editTask/{{_id}}">Edit</a>
                 </td>
             </tr>
         {{/each}}
